requirements.txt
torch
torchvision
pandas
numpy
pillow
scikit-learn

